MLHR-1858 #comment File input operator that emits the lines of the file
